adf_dir: fail operations on too long filenames, #97
So far, the operation were silently truncating the filenames that
were too long for OFS/FFS (30 characters). The client code was not
aware of that, what could result in unexpected behaviour:
- creating/renaming entries to names different than requested
- opening and reading/writing a file with a different name than
  requested (if the truncated name matched another existing file)
and all of that was silent(!).

This modification changes that behaviour - if the requested name
is too long, than it is automatically assumed the operation on it
must fail.
This returns control over what happens to the client code.