adfDevDumpOpen: fail immediately if rw not possible, #98
So far, the function was silently (with only a warning) opening
a dump device read-only as a fall-back if read-write was not
possible. The client code was not informed about it, so this could
lead to confusion if writing operations later failed (seemingly
for no obvious reason).
This modification makes the function fail immediately if write
access is requested but the dump file cannot be opened in write mode.